Minor changes in the DB.

+++ FitnessClub/Models/Subscription.cs
@@ -11,7 +11,7 @@ public partial class Subscription
 
     public string Title { get; set; } = null!;
 
-    public short? Rating { get; set; }
+    public decimal? Rating { get; set; }
 
     public virtual ICollection<Discount> Discounts { get; } = new List<Discount>();
